requirements.txt
boto3==1.37.31
botocore==1.37.31
Flask==3.1.0
Flask-Bcrypt==1.0.1
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
flatbuffers==24.3.25
graphviz==0.20.3
gunicorn==23.0.0
keras==3.7.0
matplotlib==3.9.2
mediapipe==0.10.14
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python==4.11.0.86
packaging==20.9
pandas==2.2.2
PyJWT==2.10.1
pyparsing==3.1.4
python-dotenv==1.1.0
realtime==2.4.2
requests==2.32.3
scikit-learn==1.5.2
seaborn==0.13.2
SQLAlchemy==2.0.40
StrEnum==0.4.15
supabase==2.15.0
supafunc==0.9.4
tensorflow==2.18.0
tensorflow-hub==0.12.0
tensorflow_intel==2.18.0
tensorflowjs==3.18.0
tf-estimator-nightly==2.16.0.dev2024012409
tf_keras==2.18.0
